Altera a estrutura de pastas dos adapters
