+++ grails-app/services/icaruswings/PayerService.groovy
@@ -2,7 +2,7 @@ package icaruswings
 
 import grails.gorm.transactions.Transactional
 import grails.validation.ValidationException
-import icaruswings.utils.adapters.PayerAdapter
+import icaruswings.adapters.PayerAdapter
 import icaruswings.utils.repositories.PayerRepository
 import icaruswings.utils.validator.PostalCodeValidator
 import icaruswings.utils.validator.ValidateCpfCnpj
